Refactor: migrate to cents-based amounts & strict daily pool
Major refactoring to improve precision and testability:

- Migrate all monetary amounts from Double to Long (stored in cents) to avoid floating-point precision issues
- Update Room database schema (v1 → v2) with migration to convert existing amounts
- Inject Clock dependency into MainViewModel for deterministic testing instead of systemTimeOverride
- Extract budget calculation logic into new BudgetCalculatorUseCase domain class
- Change budget logic from continuous amortization to strict daily pool: dailyLimit now based on spending before today only, not intra-day transactions
- Add CurrencyUtils.parseAmountToCents() and formatAmountInput() for safe input handling
- Migrate BudgetPreferences to use Long with legacy Double migration support
- Update all UI components, dialogs, and tests to work with cents-based amounts
- Remove unnecessary transaction sorting from UI (already sorted in ViewModel)

+++ app/src/main/java/com/imanhaikal/memo/data/Transaction.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 data class Transaction(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,
-    val amount: Double,
+    val amount: Long, // Stored in cents
     val note: String,
     val date: Long // Epoch millis
 )
